Include blacklisting option in the analysis job generation
Also add another check to check_broken.cc which scans over the directory tree and finds missing tree numbers for each sample. Basically, our Python scripts assume that if the sample is distributed into N root files, then all the root files are present by following this numbering convention: tree_1.root, tree_2.root, ... tree_N.root. However, it might be the case that the some files are missing from the remote site. For example, if there are N files, the last file must not always be tree_N.root but can also be tree_M.root (M > N). This must be taken into account. So the change in check_broken.cc is: find all missing root files between 1 and M and form the so-called blacklist of indices. The blacklist should be included in the Python configuration for each sample separately. This ensures that we always chain present files together whenever we run the analysis (and that every job processes equal number of files).