Fix symlink problem in copying the root files over to scratch
The problem was that when the sbatchManager copied the files over to /scratch, it copied only the symlink data/FR_tau_2016.root; however, if the jobs is running on a different node it probably doesn't see the file the symlink is referring to... Or something like that. Anyways, the problem was fixed with using the dereference flag (-L) in the cp command.

Also, completed the hadd of output_root files (changed the dependency to stage1 hadd output files; probably should redirect the output_root files to /scratch path and then copy the files over once the job is complete -- would make the dependency more obvious).