Improve Ntuple filling for MEM
Restrict the number of selected jets to 2 although we might throw out valid jets that actually match to the generator level quark. This is fortunately a minuscle effect. Also drop the event if the same reconstructed jet matches with both b-quark and anti-b quark within dR < 0.5. Of course the number of events can be increased a little bit by going through a lot of combinatorics, but this is suboptimal b/c we only need a handful of events in order to debug the MEM code.