Bugfix in Ntuple filling for MEM
Apparently, the check for non-null, if(!number), fails if the number is,
like, an unsigned long long or something. I don't know. Explicit check
works and we don't throw away 30 events for no reason.
Also, do not dump all events which passed the analysis but didn't pass
generator level selection.