Force b/t-quarks on mass shell
It turns out that defining b-quarks by subtracting W boson from t-quark puts the b-quarks way off from the mass shell. It's just so bad. The hack here is to redefine b-quarks' mass and energy so that they have always the same mass (their direction is kept the same). The 2nd d.o.f is covered by the requirement that the mass of b+W always equals to the top mass (which is a constant).

Also, instead of comparing mass of the gen tau pair to gen Higgs, we should compare it to PDG value of the Higgs, because gen Higgs might also be far away from its expected value of 125 GeV. In principle, the same trick is needed here as we have in the top legs right now, but its a little more complicated...