Pull MEM 3l1tau interface object outside the evaluation loop
Since every possible integration starts w/ the same seed and the operator() doesn't alter the MEM score evaluation it's possible to speed the main loop by pulling the interface object creation outside the loop. The effect is that there won't be any overhead in initializing the object (which includes reading files from the hard drive), thus it's faster.