Fix compilation error in BDT filler
For some reason 80X was unable to compile files with NtupleFillerBDT.h included. The problem was that GCC 5.3.0 thought stuff referenced by this pointer might be unininitialized, which is ridiculous. So I had to explicitly capture a variable which is already referenced by this pointer; from the immediate context of lambda this shouldn't be needed (as seen with GCC 4.9.3). I blame the compiler ! :)