Add mem output to BDT ntuple filler of 2lss1tau analysis
Had to introduce a hack in which the jets are cleaned differently, i.e. w.r.t selected leptons (as in addMEM code) instead of fakeable leptons (as in the analysis). This discrepancy in the cleaning means that addMEM code considered different jet combinations than it was supposed to. The consequence is that we lose some events after running the analysis on addMEM output; if we use the original cleaning the analysis then some of the MEM evaluated events have either no MEM objects or the constituents of the objects (i.e. leptons) don't correspond to the objects the analysis selects.

Must be researched further..