Add current workspace to CMSSW_SEARCH_PATH
The addMEM_3l_1tau executable cannot find the MadGraph file
param_card.dat since it's residing in tthAnalysis/tthMEM project
repository which is not in CMSSW's search path. Had to explicitly add
the current workspace to the list of CMSSW's search paths. In my
experience this doesn't break anything in the analysis.