Fix bug related to signatures of (collection)selector classes
I noticed that when I issued

  RecoElectronCollectionSelectorTight tightElectronSelector(era, -1, true);

I didn't see any debug output. The reason is that RecoElectronCollectionSelectorTight initializes a private member variable selector_ of class type RecoElectronSelectorTight as such:

  selector_(era, index, debug);

even though the RecoElectronCollectionSelectorTight class has the same signature.
The problem with this is that the constructor of RecoElectronSelectorTight had the following signature:

  RecoElectronSelectorTight(int era, bool set_selection_flags = true, int index = -1, bool debug = false);

Therefore, the the selector_ variable was initialized to

  selector_(era, true, 1, false);

instead of

  selector_(era, true, -1, true);

as was originally intended. In fact, the *CollectionSelector* classes always initialized the underlying selector classes to (era, true, 0), because (int)false is 0 and (bool)-1 is true. So it looks like it didn't introduce any bugs in the analysis although it was really error-prone.

The solution is to move set_selection_flags at the very end of each *CollectionSelector* signature. Also added explicit keyword to all selector constructors so that the implicit boolean-to-int and int-to-boolean variable casting never happens again.

TODO: enable a compilation flag which notifies the user about such implicit casts.