Update decay chain parser
Discovered that the URL to remote ROOT files can be completely general, i.e. there's no need for polling available tiers and construct the URLs manually (given the list of available tiers), rather the URL can be prefixed with a global domain name. If the request to open a ROOT file is done via this global domain name, the CMS servers figure out the available tiers and attempt to open the file behind the scenes.