Add tmp dir option to dump_rle_parallel
It turns out that if an sbatch job is immediately submitted once the input shell file is written (or ,,written'') to /hdfs, the job submission might fail because the necessary file is missing its contents. My hypothesis is that Hadoop requires at least 3 redundant copies of each file and there's some delay in making these copies. The solution is to keep the temporary files (logs, py and shell cfgs) on a separate storage which doesn't employ such redundancy checks (e.g. /home), hence no delay and the job can be immediately submitted.

Just tested it and worked.