Fix issues w/ DYJetsToLL_M-50 in check_broken
The process_name must be unique in each sample (it doesn't matter if its an extended sample or not). This caused some problems regarding the analysis, e.g. ext1 samples weren't analyzed at all and most of the production Ntuples belonging to ext1 weren't analyzed.

Added some failsafes to check_broken so that this won't ever happen again.