Place hadd cfgs to /home
I suspect that due to redundancy of Hadoop (i.e. making 3 copies of each file) whenever a new (text) file is created on /hdfs, it takes a little while until the file is ,,actually there''. For example, if the Makefile starts hadding stuff, it creates a bunch of shell config files on the fly which are immediately passed to SLURM. It oftentimes complains that the passed file is empty even though it isn't after certain time has passed (which varies greately, so no time.sleep() is gonna work). Even though I have done everything to ensure that the file is actually written to the disk (fflush'ing and fsync'ing the file descriptor), it seems that this is not enough, if the config files are written to /hdfs. The solution is to create and keep all text-based files on /home.

Tested the new solution with addMEM_2lss_1tau configuration; worked. Haven't tested it with default analysis configuration, but it should work (at least the config files are generated without a hitch).