Bugfix in check_broken: nof events rounding errors
The integral of TH1 is a floating point number, but we kept adding the numbers to unsigned long long, so the floating point number were always implicitly rounded downwards. This introduced small differences in event counts when compared to the actual nof events (up to 0.025% in some samples). Although the effect was really small, it's better to do it right b/c otherwise it would make comparing the dictionaries of production and addMEM Ntuples a bit annoying.

Also, discovered, that std::regex exits; made parsing the names of root files more restrictive.

(:100:0)