Temporarily remove heppyFlavor from RecoJetReader
Added it to RecoJetWriter, though, so that it can be re-enabled in RecoJetReader again.
This wasn't a problem before commit https://github.com/HEP-KBFI/tth-htt/commit/7526f5d1816fdaa00da52d09302a563795fe7552 because the error level was too high for abandoning the program execution and just some error messages were printed on the screen. The errors applied to production and addMEM Ntuples only (since RecoJetWriter created the jet branches which we read in RecoJetReader).