Do not check the file sizes immediately
It might be the case that right after file has been copied its size appears to be zero. In order to counteract it, we poll stdout and stderr of fuser command and see whether there are any processes still using the file. If not, it should be ready to use for everyone.

There's still exists a possibility that OS buffers the data somewhere and hasn't actually written it to a disk (and thus bypassing fuser). Needs more testing in the wild, though.