Refactor 1l2tau channel's code to include new TTreeWrapper
However, there seems to be some sort of race happening during the exit when THDFSFile is being used instead of regular TFile. The program runs fine, but when it tries tries to clean up at the very end (main() literally going out of scope), then it looks like multiple destructions in various threads try to free the same memory region and crashes before the main() function could return 0. This does not affect the runtime behavior or our analysis results, but would require us to parse the error code (to see if it equals to 134 as this type of error manifests itself as SIGABRT), but that would not guarantee that this exit code was raised by some other problem.

Disabled the part where a path starting with /hdfs is transformed into hdfs:///rest/of/the/path, because this would cause errors in other analyses and for everyone.

This refactoring becomes questionable; we could win with this approach only if we don't request the total number of events at once but instead we accumulate the total nof events in the event loop. This would mean that we won't open all files at once per one analysis job (unlike what TChain does) and therefore would save us some IO-headache.

Thirdly, a proper workaround to all this if forget about THDFSFile and just blatantly copy files from /hdfs to /tmpfs in the compute node, do work there and then copy the results back to /hdfs. All copying should be done via the hdfs stack instead of POSIX commands. Will be done in the near future.