Add a disclaimer to THDFSFile
Basically, when one decides to still use THDFSFile but encounters some errors when the program terminates, in particular the error:

*** glibc detected *** XXXXXX: corrupted double-linked list

it can be suppressed by setting the environment variable MALLOC_CHECK_ to 0 with

export MALLOC_CHECK_=0

However, this will only mask possible memory corruption happening at the exit of the program. My suspicion is that these errors happen because of our incomplete implementation of the THDFSFile class (remember that ROOT build of the same class is much more complex as it adds additional definition-declarations to the class which we are currently missing; re-enabling them does not fix the error, though). Since ROOT performs some magical memory de/allocation behind the scenes, I suspect that static and global destructors race with each other and therefore cause a double-free somewhere. Since this is happening at the same time as returning from main(), any program using the THDFSFile will crash while terminating.

When disabling such double-frees by setting MALLOC_CHECK_ to 0, such heap corruption errors won't be displayed and the exit code remains 0 (instead of 134 or SIGABRT) as it should. However, this could also mask errors made by ourselves. Alternatively, we could enable this malloc check just before returning from main, but it would still be a hack.

The errors may also arise from the differences in architecture for which the libjvm and libhdfs are compiled against.