Read gen matched objects instead of building them all over again
Also, there was an issue related to conversion GenHadTau -> GenParticle -> GenLepton, in which the pdg Id of the tau was set to 0. This caused errors because the particle charge was computed from the pdg Id. Changed the computation of the charge, as well as provided more detailed constructor for GenLepton that also accepts the charge.