Revert GenLeptonReader
It might be the case that the pdg Id and charge are contradictory. Since the pdg Id ,,stores'' more information about a given particle, we decided to have it a preference over the charge. In other words, we'll still compute the charge from the pdg Id. Let's still keep the augmented constructor for GenLepton, though.