Correctly interpret missing gen matched particles
If there is no generator-level match for a particle, we write a null particle with 0 4-momenta as the match. Previously, we wrote the particle through LorentzVector class (and also read it back via the same class) which set the particle's pT to infinity. When we read the gen-matched particle back, we assumed that the pT of a null particle remains zero, but apparently this is not the case due to internal intricacies of the LorentzVector class. This in turn meant that if we pre-computed the generator level matches and then ran the analyses, we would use these null particles as true gen-level matched particles. This caused different categorization of the selected events by their gen matches, different event yields and distributions in terms of discriminating variables.

The fix is to directly use the particle's pT in order to assess if the particle is null or not (the isValid_ flag in the Partcle class). This fixed the event yields and distributions in our new workflow we're about to switch to.