Update and streamline addMEM workflow
The addMEM workflow now takes the ,,slim'' production Ntuples as an input. We really don't need to apply a run-lumi-event selection to these Ntuples because we've already computed the number of MEM combinations beforehand. If the event doesn't pass a given lepton and tau selection, the number of MEM combinations is set to zero and no integration is performed. Therefore, the number of jobs shouldn't increase because we would have to run through every event anyways. Only the time it takes to compute the MEM scores increases due to little overhead between GetEntry and checking the maxPermutations_addMEM_* branches. The latter is automatically constructed from the lepton and tau selections, and is threfore completely streamlined (no human intervention needed).

Succeeded in generating the config files. Tested the executable with once config file. Haven't tried to submit the jobs to SLURM, though.