Do not mix LD_LIBRARY_PATH and PATH variables w/ their older versions
When we want to hadd with the cache settings (in order to speed up the hadding), we have to use 94x or newer CMSSW software stack. Since we don't want to create new working space nor mix older libs and binaries with the libs & binaries used during the hadding step, we have to explicitly set the environment variables ourselves and restore them back to normal once hadding is done.