Template-base branch address setting
Instead of using a function for each primitive type (int, float, their arrays etc) to set the branch addresses for the purpose of filling the branches, we can do a lot better with a single template that is able to deduce the correct string format by looking at the data type itself. Avoid potential compilation errors when the source data type and the desired one (specified by the function name) do not match.