Rename MET branches when reading
Renamed the following:
- shifted_JetEnUp & shifted_JetEnDown -> jesTotalUp & jesTotalDown
- shifted_JetResUp & shifted_JetResDown -> jerUp & jerDown
- shifted_UnclusteredEnUp, shifted_UnclusteredEnDown -> unclustEnUp & unclustEnDown

Also, the 4-momentum component is not at the end of the branch names, but in the middle; i.e. previously we had <MET>_<systematics>_<component>, but now it's <MET>_<component>_<systematics>, e.g. MET_jerUp_pt instead of MET_pt_jerUp. If the branch base name for the covariance matrix is not provided, the base name of the 4-momentum is used (instead of defaulting to ,,met'', which is hardcoding the branch name).