Do not use all available threads when running XGB
When XGB is initialized, it tries to allocate as many threads as there are cores in the system. Fortunately, this behavior is configurable, but the documation nor the API isn't clear about it. Setting the parameters explicitly on the unpickled data variable didn't work. It turns out that the parallelization is achieved via OMP (as described here: https://github.com/dmlc/xgboost/issues/1345#issuecomment-231898237). Thus, the workaround was to set the number of available threads to sklearn by setting up an environment variable. This variable won't be seen from the shell session which launches the application, b/c the whole process will be executed in a fork of the original shell.