Correct usage of idMVA_dR03 bitmask in tau selection
The values stored in idMVAoldDMdR03 (binary represenation in parenthesis) are:
- 0 (0) -- no pass
- 1 (1) -- VLoose
- 3 (11) -- Loose
- 7 (111) -- Medium
- 15 (1111) -- Tight
- 31 (11111)-- VTight
- 63 (111111) -- VVTight

Since we want to add VVLoose WP, which is computed for taus that don't pass VLoose, i.e. that have 0 ID, we need to shift the positive IDs to the left and add 1:

(ID << 1) + 1

For instance, 11 becomes 111 and 111111 becomes 1111111. When we write it back, we shift the positive IDs back to right. Note that when shifting VVLoose (1), it goes back to 0 naturally:

ID >> 1

The cuts on the WPs remain the same:
- VVLoose -- 1
- VLoose -- 2
- Loose -- 3
- Medium -- 4
- Tight -- 5
- VTight -- 6
- VVTight -- 7

The cut can also be a negative number or zero which means that all events can pass the WP selection. In order to compare the ID to a WP, we need to shift the latter to the left and perform logical AND operation:

Medium 111 shifted left + 1 -> 1111
Medium WP (4) minus 1 (3) and shifted left -> 1 << (4 - 1) is 1 << 3 is 1000
Therefore: 1111 & (1 << (4 - 1)) is 1111 & 1000 is 1000, which translates to ,,true'' boolean value
If ID is Loose, then it would fail (111 & 1000 = 0), but IDs greater than Medium still pass (e.g. VTight: 111111 & 100000 = 100000).