Possible to keep the files of individual jobs on /home
Now that /home is buffered by SSDs, we are able to create and keep the files of individual SLURM jobs on /home instead of /hdfs. This should increase the responsiveness of /hdfs (since /hdfs and /scratch sit on the same physical drive), and the jobs should run faster because the SSDs (/home) are more capable than spinners (/scratch) in terms of iops.

Tested with the /home option, but still defaulting to /scratch for the moment. Haven't tested the default option, though.