Check if ROOT tries to recover an output file during job resubmission
The problem is that if ROOT tries to recover a file, it doesn't set any error flags. Therefore, setting the abortion level to a very low value has no effect in detecting whether ROOT tries to recover the file or not. The alternative is to temporarily redirect the stdout and stderr to a variable. However, it has its own problems too: the destination must be a file (descriptor). We don't want to create a new file every time we check if the output file (which already exists) is corrupted or not. But creating a temporary file with a constant name doesn't work either when two users (or the same user) try to resubmit two different analyses at the same time. Of course, we could pass a unique version to the generate_sbatch_line() function, but things get messy as we would need to edit 100s of other lines in 10s of files.

The only managable solution is to check the recovery information via a bash call. The only downside here is that we do the same work (i.e. opening the same file) twice.

Untested.