Generate dictionary from the original nanoAOD Ntuples
The problem was that the nanoAOD files do not include the histograms which indicate how many events or how many weighted events this particular ROOT file contains. Even though we don't need this information in the first iteration of the analyses, i.e. at the Ntuple production step, we still need to generate dictionary that is fed to the python config generator for the Ntuple production.

Thus, if the histograms are not available, nothing won't break down. Additionally, we read the number of TTree events which is always be available (otherwise there is no analysis). Thus, we record three types of event counts:
1. Number of events listed in the DAS DB -- pre-filled by a human in the meta-dictionary;
2. Number of events available in the ROOT file -- always possible to fetch this information. Also, we probably are gonna switch to this figure instead of the weighted event counts as the latter isn't that accurate in splitting e.g. addMEM jobs;
3. Number of weighted events -- read from a histogram generated during the Ntuple production step, i.e. it's available only if we've run produceNtuple.sh. This figure is used in luminosity scaling.