Version the log files
Previously we have versioned only the stdout/stderr of the make command, but not the log files of individual jobs -- they were overwritten every single time the jobs got resubmitted. From hereon out, all log files have versions attached to them, so that it would be possible to track down potential bugs in the workflow. Tested for Ntuple production jobs.