Update analysis-specific trigger selection
In case of 0l2tau and 0l3tau the triggers are the same as specified in analyzeConfig_new.py; but for 1l1tau and 1l2tau the trigger selection is a bit more narrow. Unfortunately, the triggers specified in these configs are not available for 2017 analysis and I haven't come up with better alternatives, yet.