Update crabSitter.py
Instead of hardcoding the paths to the script, I prefer to use command line parameters where possible. The only thing that was cumbersome to pass via CLI is the list of samples (or SAMPLES in the script) that are only checked.

Also, it was a mystery to me why there were two CRAB directories... I believe that one was used to save the JSON file (under Heppy/test) and the other was the actual directory which contained the CRAB log files (on /scratch). I didn't spot any dependencies on the Heppy code/configs, so hopefully there is none.