Refactor writer classes
Previously, when the setBranches() function was called, it first initialized the arrays and then set the branch addresses, e.g. in pseudocode

someVariable = new Int_t[someSize];
tree -> Branch(someVariable, someName, someNameForROOT)

This, however has two problems:
1) when the type of some variable changes, one has to change the both lines: the type after 'new' and 'someNameForROOT';
2) when trying to add one more variable (or a new writer class), a human error may occur when a person forgets to modify the copied lines (e.g. the branch names or variable names);

In order to mitigate these types of errors, I found a solution which initializes the variables in proper types and sets the branch names, all in one go. When someone needs to change the type of a member variable of some writer class, then all that needs to be done is to change the type in the class definition. The syntax reduces to:

bai.setBranch(someVariable, someName)

where bai is an instance of BranchAddressInitializer -- the class that does the heavy lifting of initializing the arrays and setting the branches.

Also cleaned up some code: used range-for loops where possible, removed unncessary header inclusion statements.