Write executable log files straight to the assumed destination
If the job directory resides on /home, then we just duplicate the log files on the storage, which is counterproductive. Besides, if there's something preventing to copy the log file, the job get stuck for indefinite amount of time and might die. We write the wrapper files to /home anyways.

Used both /home and /scratch as the job directory in testing.