Remove recording the branch names in the reader classes
The initial idea what motivated me to implement this feature in the first place was that when we create new Ntuples (with produceNtuple or the addMEM programs) by copying the branches from the input Ntuples, we don't have to keep track of the branch names which should be excluded in order to avoid double-reading. However, it just occured to me that we cannot do this for arrays which we don't read with the reader classes but their length variable we read with the reader classes. This is especially problematic when the input and output branch names are identical (which we want to retain). For instance, say we branches

Electron_pt[nElectron];
Electron_jetIdx[nElectron];

The first one we read with the reader class, including the nElectron variable, but the second we don't. Now, the initial idea implied that we can simply ignore branch names nElectron and Electron_pt because they're reconstructed by the reader-writer classes, and that we can directly copy over Electron_jetIdx. But Electron_jetIdx depends on the nElectron variable, which has a different value in the output Ntuple compared to the input Ntuple. Therefore we either have to read all possible Electron branches or abandon the idea altogether. I chose the latter since reading all branches is sub-optimal.

There's a workaround to this (which will be implemented very soon): we can ask the regex from the reader classes to rule out all Electron (or any other collection) classes. This still has the advantage that the user doesn't have to specify the branch names twice in the python config file (possibly introducing errors).