Revert back to simple TChain approach in produceNtuple
Since the timescale is tight and flu is all over the place, I'll revert the attempts to use TTreeWrapper in produceNtuple. The main problem is that it's non-trivial to copy the input branches from multiple Ntuples to a single output Ntuple with TTreeWrapper. Possible solution probably is to embed the logic of copying the branches into the TTreeWrapper class, but this requires some investigation first.