Drop gen-matched object multiplicity branches in produceNtuple
The benchmark for validating produceNtuple code is to:

1) produce one Ntuple from post-processed Ntuple: original.root -> first.root
2) take this output as an input and run produceNtuple on it: first.root -> second.root

If first.root and second.root are not identical, then there's a problem in produceNtuple. The current problem is that we didn't drop the input branches computed from previous iteration of produceNtuple, namely the number of gen-matched objects. This doesn't have any major consequences to the analysis because we run produceNtuple only once in practice. Nevertheless, it's nice to know where the discrepancy comes from.