Provide alternative constructors in Reader/Writer classes
In the alternative constructors there's no need to specify the branch name for the multiplicity variable of a given particle collection. In my experience, the multiplicity branch is always of the form n<collection name>, thus specifying this branch name is completely redundant and can be omitted in the code. This will be soon done in the next commits.

Only GenLeptonReader is currently missing the alternative prescription since the constructor expects two collection names and there would be ambiguities which constructor to use. Will likely drop the branch names for leptons from tau since the gen lepton branch already contains all generator-level leptons.