Check for duplicate branches after detecting branch type
Otherwise, the checks and corresponding warning messages will be emitted prematurely. Previously, if the singular-type branches are copied over first, followed by checks on vector-type branches, the latter would complain about every single-type branch being a duplicate. Since the probability for having duplicate branches in a ROOT file is really low, it's more optimal to move this check further down the for-loop.