Do not figure out the btag SF branch name in executable code
Previously, each executable did two things: parsed the systematic uncertainty string to figure out, if the given sys uncert was relevant to b-tagging SFs; if it was, then the code in each executable figured out the corresponding branch name through some auxiliary code. This branch name was then passed to RecoJetReader, which in turn parsed the string again to substitute the correct jet collection name. This kind of indirection is totally unncessary. If there's ever a need to change the b-tagging SF branch names, one has to go through each and every executable and rename the branch. Plus there's potential for errors in the way the RecoJetReader/Writer classes parse the passed branch name.

Currently, each executable code figures out the corresponding b-tagging SF enum, which is passed to RecoJetReader. Both RecoJetReader and RecoJetWriter use the enums to figure out the correct branch name using an auxiliary function. If there's a need to change the b-tagging SF branch names, only this aux function must be updated; the remaining code stays untouched.

Compiles, but haven't tested, yet.