Integrate SyncNtupleManager into 1l2tau, 2lss1tau & 3l1tau channels
The synchronization Ntuples are produced only if both the output TTree name and the path to the output file are specified. By default, the basis cfg files do not specify these values, which means that SyncNtupleManager is totally harmless in the analysis executables.

Soon the sync_ntuple executable will be depracted since it's counter-productive to maintain yet another executable: if there's some kind of analysis-specific change, it has to be propagated to sync_ntuple. Besides, keeping a separate executable doesn't help in streamlining sync Ntuple production at all.