Integrate sync Ntuple production into 1l_2tau analysis workflow
We basically use the sync Ntuple as the output file, and not the hadd stage 1 files. Needed to write auxiliary functions to analyzeConfig that adds relevant lines to the Makefile (replacing hadd stage1 files with the sync Ntuples and hadding the Ntuples). Also removed era-specific trigger selection from the 1l_2tau config since this issue was relevant in 2016 analysis (e+tau cross-triggers were missing at that time).