Automatically decide which MC samples belong to the same phase space
While generating the meta-dictionary, it is possible to find out the samples which belong to the same phase space if the naming conventions are kept intact. The decision is made based on the so-called specific names of the samples. For instance, samples ZZTo4L and ZZTo4L_ext1 cover the same phase space and therefore their cross sections are indentical. Same holds for PSweights and non-PSweights samples (e.g. TTTo2L2Nu and TTTo2L2Nu_PSweights). This information (i.e. the set of samples that cover the same phase space) is used in create_dictionary step, which sums the weighted event counts and sets the event counts to the calculated sum for all the samples belonging to the same phase space.