Add more information to the sample dictionaries w/ create_dictionary
Added the following information:
- number of DB files -- easier to compare the nof MINIAOD(SIM) files against the numbers of nanoAOD Ntuples
- total file size in the sample -- may be useful when trying to estimate potential usage of the storage space
- total file size of the MINIAOD(SIM) files -- to get a simple reduction figure on the file sizes

The file sizes are integer bytes (which may help us to dynamically find the optimal number of produceNtuple files that contain pre-selected objects which we can concatenate), but the human-readable figures are in the comments.
Also, removed the full path to the create_dictionary.py from the comments so that people can directly copy the whole line when re-generating the sample dictionary.
Had to move human_size() function to jobTools; figured that we can re-use run_cmd from the same module in find_samples.py as well.