Match up to most one gen level particle to a reco particle
In other words, there can be either one gen match for one reco particle, or none at all. Also added a feature which requires a gen particle to be matched up to one reco particle (so that the gen particle won't be matched to two or more different reco particles), but this feature is disabled for now.