Let RecoHadTauCollectionSelector inherit from ParticleCollectionSelector
The code in RecoHadTauCollectionSelector is identical to ParticleCollectionSelector; the former just adds some functions regulating the selector_ member variable. This change reduces boilerplate code and pulls the selector classes towards the central ParticleCollectionSelector class. In other words, if there's a need for a fundamental change in ParticleCollectionSelector, it's likely that this needs to be propagated to RecoHadTauCollectionSelector as well. Inheriting from ParticleCollectionSelector solves this potential problem automatically.