Remove pT cut in b-tagging selector at produceNtuple step
Otherwise, there would be absolutely no point in selecting the cleaned jets based on the condition that their pT in any JEC setting passes the preselection, because the central pT would still be used in loose and medium b-tagging selectors; the maximum possible number of permutations is estimated from the multiplicity of loose and medium b-jets. The correct way of doing this would probably be computing the max nof permutations for each systematic uncertainty, but this would mean that there would be 7 times as many branches needed. In practice this number barely changes; it's better to overestimate the nof permutations because when we actually compute the MEM scores for each permutation of input objects, because in the case that max nof permutations is exceeded, the code can throw or just neglect further permutations.

So far we've only considered the effect of JEC on the nof permutations; other systematic uncertainties should also be considered to get the true upper bound on the nof permutations (per channel; lepton and tau definition; tau WP).