Add sync_noMEM mode to all analyses subject to the sync
In the first approximation, we really don't need to compute the MEM scores to see, if we've have improved in other variables. Once we've achieved a desireable level of synchronization in non-MEM variables, we can compute the MEM variables afterwords. (The MEM step is realtively slow -- 1-2h -- compared to running the analyses on the sync Ntuple, which takes roughly 10-15 mins).