Apply b-tagging selection on selJets, not cleanedJets in produceNtuple
The commit 7d5249509bd7d56847b2228ea302c6a558076c73 was incomplete: even though we removed the pT cut on b-tagging selectors, we passed cleaned jets to these selectors, not selJets. This was wrong b/c the b-jets, the multiplicity of which is used to compute the maximum nof MEM object permutations, were not filtered by their pT. Therefore, the estimated nof permutations was higher than it actually was.

A few tweaks: MEMPermutationWriter::find_selection_str is now a static function; added isDEBUG flag to all selectors and cleaners in produceNtuple; single-character string literal changed to a character.