Read/write JES/JER branches in RecoMEt/JetReader/Writer + refactor
This commit includes multiple changes which are all tied together (but hopefully the modularity improves a bit):
- read & write jet pT and mass, and MET pT and phi based on the central_or_shift option passed by the user;
- also, added the capability to read & write non-nominal versions of these variables, either for the sync or for the data;
- decided to migrate all enums that represent the systematic uncertanty to a new place called sysUncertOptions;
- this file also contain the functions that convert any given sys uncertainty option from string to enum, and also auxiliary functions to construct the branch names in RecoMEt/JetReader/Writer classes;
- the former functions helped to remove boilerplate code from individual executables;
- LHE weights are now read with a single line in individual executables;
- never write any branches that represent systematic uncertainties to data Ntuples.

Tested with produceNtuple in three regimes:
- isMC true, useNonNominal true -> non-nominal (i.e. Jet_pt, Jet_mass, MET_pt, MET_phi) are used;
- isMC true, useNonNominal false -> nominal (i.e. Jet_pt_nom, Jet_mass_nom, MET_pt_nom, MET_phi_nom) are used;
- isMC false useNonNominal true -> non-nominal are used;

If isMC is false and useNonNominal is also false, then the code throws (as it's supposed to) because the data doesn't have nominal (i.e. MET-adjusted) MET and jet branches.

From now on, if saving of the systematic uncertainty branches is enabled, then non-nominal branches will also be saved (to reduce the complexity of the code). When it's no obvious, which branches produceNtuple used in the event selection, one can check whatever branches were written to the output files first -- if nominal are first, then these were used in event selection.

Testing in addMEM and analyze executables will be done shortly.