Add option to prepareDatacards to perfore quantile rebinning on fakes
instead of the whole background. Compiles, untested.
Also specified explicitly the option in 2lss1tau, 2l2tau and 3l1tau channels.
Removed apply_quantile_rebinning because it's redundant: it suffices to know the number of bins requested -- if the number is positive, then quantile rebinning is performed, otherwise it's not. For this reason set the default nof bins used in quantile rebinning from 4 to -1 so that the quantile rebinning is disabled by default.