Option to filter samples at cmd line in the test scripts
Since I've got tired of filtering out certain samples by hand either for testing reasons or in legitimate situations (such as producing Ntuples only for the signal sample), I decided to add yet another option to the scripts in test/: -f/--filter <condition>, where condition consists of two parts separated by a colon. The first part is either the sample category ('cat') or the specific (human-readable) name of the sample ('name'); the second part is regex-formatted string that should filter out the samples on which the user wants to run the analysis. Examples:

* -f cat:signal -- run the analysis on the signal samples only
* -f cat:TTZ|TTW -- run the analysis on TTW and TTZ samples only
* -f "name:*fastsim*" -- run the analysis on FastSim samples

Notice that the filter option overwrites the default sample selection that is determined by the -m/--mode option; -f/--filter should only be used in special circumstances.